+++ apps/backend/src/libs/types/repository.type.ts
@@ -3,7 +3,7 @@ type Repository<T = unknown> = {
 	delete(id: number, userId: number): Promise<boolean>;
 	find(id: number, userId: number): Promise<T | null>;
 	findAll(id: number): Promise<T[]>;
-	update(id: number): Promise<T | null>;
+	update(id: number, payload: unknown): Promise<T | null>;
 };
 
 export { type Repository };