+++ apps/frontend/src/libs/constants/menu-items.constants.ts
@@ -1,6 +1,6 @@
 import { AppRoute } from "~/libs/enums/enums.js";
 
-import { MenuItem } from "../types/types.js";
+import { type MenuItem } from "../types/types.js";
 
 const MENU_ITEMS: MenuItem[] = [
 	{