+++ packages/shared/src/modules/friends/libs/exceptions/friend-error/friend-error.exception.ts
@@ -1,5 +1,8 @@
-import { HTTPCode, HTTPError } from "../../../../../libs/modules/http/http.js";
-import { ValueOf } from "../../../../../libs/types/value-of.type.js";
+import {
+	type HTTPCode,
+	HTTPError,
+} from "../../../../../libs/modules/http/http.js";
+import { type ValueOf } from "../../../../../libs/types/value-of.type.js";
 
 class FriendError extends HTTPError {
 	public constructor(message: string, status: ValueOf<typeof HTTPCode>) {