+++ packages/shared/src/modules/friends/libs/types/types.ts
@@ -1,3 +1,2 @@
 export { type FriendDto } from "./friend-dto.type.js";
-
 export { type FriendFollowingRequestDto } from "./friend-following-request-dto.type.js";