+++ packages/shared/src/modules/users/users.ts
@@ -1,12 +1,15 @@
+export { UsersApiPath } from "./libs/enums/enums.js";
 export {
 	type UserAuthResponseDto,
 	type UserGetAllResponseDto,
+	type UserProfileRequestDto,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
 } from "./libs/types/types.js";
 export {
+	userProfile as userProfileValidationSchema,
 	userSignIn as userSignInValidationSchema,
 	userSignUp as userSignUpValidationSchema,
 } from "./libs/validation-schemas/validation-schemas.js";